# docs/swagger.yaml
basePath: /
definitions:
...
  merchant.CreateServiceRequest:
    properties:
      category_id:
        type: integer
      cover_image:
        type: string
      description:
        type: string
      duration:
        type: integer
      name:
        type: string
      price:
        type: integer
    required:
    - category_id
    - duration
    - name
    - price
    type: object
...
  merchant.RefundRequest:
    properties:
      refundAmount:
        description: 退款金额(分)
        minimum: 1
        type: integer
      refundReason:
        description: 退款原因
        type: string
    required:
    - refundAmount
    type: object
...
  merchant.UpdateServiceRequest:
    properties:
      cover_image:
        type: string
      description:
        type: string
      duration:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
    type: object
...
  models.Payment:
    properties:
      amount:
        description: 支付金额(分)
        type: integer
      appointmentId:
        description: 关联预约ID
        type: integer
      createdAt:
        type: string
      customerId:
        description: 用户ID
        type: integer
      description:
        description: 支付描述
        type: string
      failReason:
        description: 失败原因
        type: string
      id:
        type: integer
      merchantId:
        description: 商家ID
        type: integer
      outTradeNo:
        description: 商户订单号
        type: string
      paidAt:
        description: 支付时间
        type: string
      status:
        description: 支付状态
        type: string
      transactionId:
        description: 微信交易号
        type: string
      updatedAt:
        type: string
    type: object
  models.Refund:
    properties:
      amount:
        description: 退款金额(分)
        type: integer
      appointmentId:
        description: 关联预约ID
        type: integer
      createdAt:
        type: string
      failReason:
        description: 失败原因
        type: string
      id:
        type: integer
      outRefundNo:
        description: 商户退款单号
        type: string
      paymentId:
        description: 支付ID
        type: integer
      reason:
        description: 退款原因
        type: string
      refundId:
        description: 微信退款单号
        type: string
      refundedAt:
        description: 退款时间
        type: string
      status:
        description: 退款状态
        type: string
      updatedAt:
        type: string
    type: object
  models.Service:
    properties:
      categoryID:
        type: integer
      coverImage:
        type: string
      createdAt:
        type: string
      description:
        type: string
      duration:
        description: 分钟
        type: integer
      id:
        type: integer
      isActive:
        type: boolean
      merchantID:
        type: integer
      name:
        type: string
      price:
        type: integer
      sort:
        type: integer
      updatedAt:
        type: string
    type: object
...
  utils.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      page:
        type: integer
      pageSize:
        type: integer
      success:
        type: boolean
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 预约系统API文档
  title: Appointment System API
  version: "1.0"
paths:
...
  /api/merchant/appointments/{appointmentId}/refund:
    post:
      consumes:
      - application/json
      description: 为已支付的预约发起退款
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 预约ID
        in: path
        name: appointmentId
        required: true
        type: integer
      - description: 退款请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/merchant.RefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 退款记录
          schema:
            $ref: '#/definitions/models.Refund'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 退款失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 发起退款
      tags:
      - 商家支付
...
  /api/merchant/payments:
    get:
      consumes:
      - application/json
      description: 查询商家的支付记录
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 支付状态
        enum:
        - pending
        - succeeded
        - refunded
        - failed
        in: query
        name: status
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 支付记录列表
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Payment'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取商家支付记录
      tags:
      - 商家支付
  /api/merchant/payments/{paymentId}:
    get:
      consumes:
      - application/json
      description: 查询支付订单详细信息
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 支付订单ID
        in: path
        name: paymentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 支付订单详情
          schema:
            $ref: '#/definitions/models.Payment'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 支付订单不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取支付详情
      tags:
      - 商家支付
  /api/merchant/services:
    get:
      consumes:
      - application/json
      description: 根据分类ID获取服务列表
      parameters:
      - description: 分类ID
        in: query
        name: category_id
        required: true
        type: integer
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回服务列表
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Service'
                  type: array
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取商户服务列表
      tags:
      - 服务管理
    post:
      consumes:
      - application/json
      description: 商家创建新的服务项目（需要商家认证）
      parameters:
      - default: Bearer <token>
        description: BearerToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: 服务信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/merchant.CreateServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Service'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建新的商家服务
      tags:
      - Merchant Services
  /api/merchant/services/{serviceId}:
    delete:
      consumes:
      - application/json
      description: 删除商家服务（软删除或硬删除）
      parameters:
      - description: 服务ID
        example: 123
        in: path
        name: serviceId
        required: true
        type: integer
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 无效的服务ID
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 服务不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 删除服务失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除服务
      tags:
      - 商家服务管理
    put:
      consumes:
      - application/json
      description: 更新商家服务的详细信息
      parameters:
      - description: 服务ID
        in: path
        name: serviceId
        required: true
        type: integer
      - description: 服务更新请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/merchant.UpdateServiceRequest'
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 无效的服务ID | 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 认证失败
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 服务不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 更新服务失败
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新服务信息
      tags:
      - 商家服务管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"